Small DBL clean up - group equipment and creature related repositories together (less spam) Player - implemented health/armor/power calculations by Blumster that take into account player race.
