+++ src/Rasa.DBL/Repositories/World/IWorldUnitOfWork.cs
@@ -4,23 +4,10 @@
 
     public interface IWorldUnitOfWork : IUnitOfWork
     {
-        IArmorClassRepository ArmorClasses { get; }
+        IEquipmentRepository Equipment { get; }
         ICreatureRepository Creatures { get; }
-        ICreatureActionRepository CreatureActions { get; }
-        ICreatureAppearanceRepository CreatureAppearances { get; }
-        ICreatureStatRepository CreatureStats { get; }
         IEntityClassRepository EntityClasses { get; }
-        IEquipableClassRepository EquipableClasses { get; }
         IFootlockerRepository Footlockers { get; }
-        IItemClassRepository ItemClasses { get; }
-        IItemTemplateArmorRepository ItemTemplateArmors { get; }
-        IItemTemplateItemClassRepository ItemTemplateItemClasses { get; }
-        IItemTemplateRepository ItemTemplates { get; }
-        IItemTemplateRequirementRepository ItemTemplateRequirements { get; }
-        IItemTemplateRequirementRaceRepository ItemTemplateRequirementRaces { get; }
-        IItemTemplateRequirementSkillRepository ItemTemplateRequirementSkills { get; }
-        IItemTemplateResistanceRepository ItemTemplateResistances { get; }
-        IItemTemplateWeaponRepository ItemTemplateWeapons { get; }
         ILogosRepository Logoses { get; }
         IMapInfoRepository MapInfos { get; }
         INpcMissionRepository NpcMissions { get; }
@@ -29,8 +16,5 @@
         IPlayerRandomNameRepository RandomNames { get; }
         ISpawnpoolRepository Spawnpools { get; }
         ITeleporterRepository Teleporters { get; }
-        IVendorItemRepository VendorItems { get; }
-        IVendorReposiotry Vendors { get; }
-        IWeaponClassRepository WeaponClasses { get; }
     }
 }
